backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
requests==2.31.0
opencv-python==4.8.1.78
deepface==0.0.98
numpy==1.23.5
apscheduler==3.10.4
tensorflow==2.12.0
pymongo==4.6.0
motor==3.3.2
python-dotenv>=1.0.1
cloudinary==1.41.0
pinecone-client==2.2.4
Pillow>=10.0.0
